+++ src/Rocks.Analysis.IntegrationTests/MockDefinitions.cs
@@ -115,3 +115,4 @@
 [assembly: Rocks.Rock(typeof(Rocks.Analysis.IntegrationTests.Referenced.IDoPublicStuff), Rocks.BuildType.Create | Rocks.BuildType.Make)]
 [assembly: Rocks.Rock(typeof(System.Net.Http.HttpMessageHandler), Rocks.BuildType.Create)]
 [assembly: Rocks.Rock(typeof(Rocks.Analysis.IntegrationTests.ShimTestTypes.IAuditTrail<>), Rocks.BuildType.Create | Rocks.BuildType.Make)]
+[assembly: Rocks.Rock(typeof(Rocks.Analysis.IntegrationTests.RockContextTestTypes.IContextural), Rocks.BuildType.Create | Rocks.BuildType.Make)]