+++ changelog.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [9.2.0] - Not Yet Released
+
+### Added
+
+- Added `DisableVerification` flag to `RockContext` to allow developers to temporarily disable verification (issue [#377](https://github.com/JasonBock/Rocks/issues/377))
+
+### Fixed
+
+- Changed `Expectations` to not call `Verify()` if the mock threw a code generated exception (issue [#377](https://github.com/JasonBock/Rocks/issues/377))
+
 ## [9.1.0] - 2025-05-11
 
 ### Added